\name{BingSource}
\alias{BingSource}
\title{Get Feed Meta Data from Microsoft Bing Search...}
\usage{BingSource(query, n=100, count=10, appid, sources="news", market="en-US", ...)
}
\description{Get Feed Meta Data from Microsoft Bing Search}
\author{Mario Annau}
\value{WebXMLSource}
\arguments{\item{query}{character specifying query to be used to search tweets}
\item{count}{number of results per page, defaults to 100}
\item{n}{number of results, defaults to 100}
\item{appid}{appid to be specified}
\item{...}{additional query parameters, see \url{http://search.twitter.com/api/}}
}

