\name{NYTimesSource}
\alias{NYTimesSource}
\title{Get Feed Meta Data from NYTimes...}
\usage{NYTimesSource(query, n=100, count=10, appid, ...)
}
\description{Get Feed Meta Data from NYTimes}
\author{Mario Annau}
\value{WebJSONSource}
\arguments{\item{query}{character specifying query to be used to search tweets}
\item{count}{number of results per page, defaults to 10}
\item{n}{number of results, defaults to 100}
\item{appid}{id to be used for query}
\item{...}{additional query parameters, see \url{http://search.twitter.com/api/}}
}

