\name{TwitterSource}
\alias{TwitterSource}
\title{Get Feed Meta Data from Twitter...}
\usage{TwitterSource(query, count=100, n=1500, ...)
}
\description{Get Feed Meta Data from Twitter}
\author{Mario Annau}
\value{list}
\arguments{\item{query}{character specifying query to be used to search tweets}
\item{count}{number of results per page, defaults to 100}
\item{n}{number of results, defaults to 100}
\item{...}{additional query parameters, see \url{http://search.twitter.com/api/}}
}

